requirements.txt
streamlit==1.32.2
google-generativeai==0.8.4
Pillow==10.2.0
python-docx==1.1.0
python-pptx==0.6.21
PyMuPDF==1.23.6
